+++ android/src/test/kotlin/com/ritiknagar/audio_metadata_flutter/AudioMetadataFlutterPluginTest.kt
@@ -15,13 +15,13 @@ import kotlin.test.Test
 
 internal class AudioMetadataFlutterPluginTest {
     @Test
-    fun onMethodCall_getPlatformVersion_returnsExpectedValue() {
+    fun onMethodCall_unknownMethod_returnsNotImplemented() {
         val plugin = AudioMetadataFlutterPlugin()
 
-        val call = MethodCall("getPlatformVersion", null)
+        val call = MethodCall("unknownMethod", null)
         val mockResult: MethodChannel.Result = Mockito.mock(MethodChannel.Result::class.java)
         plugin.onMethodCall(call, mockResult)
 
-        Mockito.verify(mockResult).success("Android " + android.os.Build.VERSION.RELEASE)
+        Mockito.verify(mockResult).notImplemented()
     }
 }