+++ lib/audio_metadata_flutter.dart
@@ -1,8 +1,10 @@
-
 import 'audio_metadata_flutter_platform_interface.dart';
+import 'src/audio_metadata.dart';
+
+export 'src/audio_metadata.dart';
 
 class AudioMetadataFlutter {
-  Future<String?> getPlatformVersion() {
-    return AudioMetadataFlutterPlatform.instance.getPlatformVersion();
+  Future<AudioMetadata?> getMetadata(String filePath) {
+    return AudioMetadataFlutterPlatform.instance.getMetadata(filePath);
   }
 }