+++ lib/audio_metadata_flutter_method_channel.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
+import 'src/audio_metadata.dart';
 import 'audio_metadata_flutter_platform_interface.dart';
 
 /// An implementation of [AudioMetadataFlutterPlatform] that uses method channels.
@@ -10,8 +11,14 @@ class MethodChannelAudioMetadataFlutter extends AudioMetadataFlutterPlatform {
   final methodChannel = const MethodChannel('audio_metadata_flutter');
 
   @override
-  Future<String?> getPlatformVersion() async {
-    final version = await methodChannel.invokeMethod<String>('getPlatformVersion');
-    return version;
+  Future<AudioMetadata?> getMetadata(String filePath) async {
+    final result = await methodChannel.invokeMethod<Map<Object?, Object?>>(
+      'getMetadata',
+      {'filePath': filePath},
+    );
+    if (result == null) {
+      return null;
+    }
+    return AudioMetadata.fromMap(result);
   }
 }