+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 1.0.0
+* Initial release of the `audio_metadata_flutter` plugin.
+* Supports extraction of various metadata fields from local audio files on Android.
+* Provides typesafe data class `AudioMetadata` for structured access.